// src/api/generated/models/paginatedChildrenResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * ECD Backend API
 * Early Childhood Development management system API.
 * OpenAPI spec version: 1.0
 */
import type { ChildResponseDto } from './childResponseDto';

export interface PaginatedChildrenResponseDto {
  items: ChildResponseDto[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}